db change from kor_data to eng_data, add reserve_page
