--- services/reservation-detail-service/404.html
@@ -1,13 +1,0 @@
-{% extends 'base.html' %}
-{% block content %}
-<!-- 404 오류페이지 -->
-<div class="container">
-    <div class="row justify-content-center">
-        <div class="col-12 text-center">
-            <span class="display-1 d-block">404</span>
-            <div class="mb-4 lead">페이지를 찾을 수 없습니다.</div>
-            <a href="/" class="btn btn-link">홈으로 돌아가기</a>
-        </div>
-    </div>
-</div>
-{% endblock %}
