--- services/reservation-detail-service/templates/reservation_detail.html
@@ -1,37 +1,0 @@
-{% extends "services/common/templates/base.html" %}
-{% block title %}주차예약상세{% endblock %}
-
-{% block content %}
-    <h1>예약 상세 페이지</h1>
-
-    {% if reservation %}
-        <div style="text-align: center;">
-            <h2>
-                {% if reservation.reservation_status == "completed" %}
-                    예약 완료
-                {% elif reservation.reservation_status == "canceled" %}
-                    취소됨
-                {% else reservation.reservation_status == "reserved" %}
-                    예약 내역 없음
-                {% endif %}
-            </h2>
-            <table border="1" style="margin: auto;">
-                <tr>
-                    <th>주차장 ID</th>
-                    <td>{{ reservation. }}</td>
-                </tr>
-                <tr>
-                    <th>사용자 ID</th>
-                    <td>{{ reservation.user_id }}</td>
-                </tr>
-            </table>
-            <div style="margin-top: 20px;">
-                <a href="{{ url_for('edit_reservation', reservation_id=reservation.reservation_id) }}">
-                    <button>수정하기</button>
-                </a>
-            </div>
-        </div>
-    {% else %}
-        <p>예약 정보가 없습니다.</p>
-    {% endif %}
-{% endblock %}
