+++ services/reservation_detail_service/app.py
@@ -1,0 +1,6 @@
+from . import create_app
+
+app = create_app()
+
+if __name__ == "__main__":
+    app.run(debug=True)
