+++ services/reservation_detail_service/templates/reservation_detail.html
@@ -1,0 +1,56 @@
+{% extends "services/common/templates/base.html" %}
+{% block title %}주차예약상세{% endblock %}
+
+{% block content %}
+    <h1>예약 상세 페이지</h1>
+
+    {% if reservation %}
+        <div style="text-align: center;">
+            <h2>
+                {% if reservation.reservation_status == "completed" %}
+                    예약 완료
+                {% elif reservation.reservation_status == "canceled" %}
+                    취소됨
+                {% else reservation.reservation_status == "reserved" %}
+                    예약 내역 없음
+                {% endif %}
+            </h2>
+            <table border="1" style="margin: auto;">
+                <tr>
+                    <th>예약 주차장</th>
+                    <td>{{ parkingLot.parkinglot_name }}</td>
+                </tr>
+                <tr>
+                    <th>예약자</th>
+                    <td>{{ user.name }}</td>
+                </tr>
+            </table>
+            
+            <div style="margin-top: 20px;">
+                <a href="{{ url_for('reservation.edit', reservation_id=reservation.reservation_id) }}">
+                    <button>예약 수정하기</button>
+                </a>
+            </div>
+            <a href="javascript:void(0)" class="delete btn btn-sm btn-outline-secondary"
+            data-uri="{{ url_for('reservation.delete', reservation.reservation_id) }}">
+                    <button>예약 취소하기</button>
+            </a>
+        </div>
+    {% else %}
+        <p>예약 정보가 없습니다.</p>
+    {% endif %}
+{% endblock %}
+
+{% block script %}
+<script type="text/javascript">
+  const delete_elements = document.getElementsByClassName("delete");
+  Array.from(delete_elements).forEach(function (element) {
+    element.addEventListener("click", function () {
+      if (confirm("정말로 삭제하시겠습니까?")) {
+        location.href = this.dataset.uri;
+      }
+    });
+  });
+
+</script>
+{% endblock %}
