+++ services/reservation_detail_service/templates/reservation_modify.html
@@ -5,26 +5,27 @@
     <h1>주차예약 수정페이지</h1>
 
     {% if reservation %}
-        <form method="POST" action="{{ url_for('update_reservation', reservation_id=reservation.reservation_id) }}">
+      <h5 class="border-bottom my-3 py-2">
+        {{ user.reservation_set|length }}개의 예약이 있습니다.
+      </h5>
+        <form method="POST" action="{{ url_for('reservation.update', reservation_id=reservation.reservation_id) }}">
             <table border="1" style="margin: auto;">
                 <tr>
-                    <th>예약 ID</th>
-                    <td>{{ reservation.reservation_id }}</td>
+                    <th>예약자명</th>
+                    <td>{{ user.name }}</td>
                 </tr>
                 <tr>
-                    <th>주차장 ID</th>
-                    <td><input type="number" name="parkinglot_id" value="{{ reservation.parkinglot_id }}"></td>
+                    <th>예약된 주차장</th>
+                    <td><input type="number" name="parkinglot_id" value="{{ parkingLot.parkinglot_name }}"></td>
                 </tr>
                 <tr>
                     <th>사용자 ID</th>
                     <td><input type="number" name="user_id" value="{{ reservation.user_id }}"></td>
                 </tr>
-                
             </table>
-
             <div style="margin-top: 20px; text-align: center;">
                 <button type="submit">저장</button>
-                <a href="{{ url_for('reservation_detail', reservation_id=reservation.reservation_id) }}">
+                <a href="{{ url_for('reservation.save', reservation_id=reservation.reservation_id) }}">
                     <button type="button">취소</button>
                 </a>
             </div>
